YAPF formatting, consolidated tests, addressed comments
